Extract code from main into AutocompleteHandler
