--- include/AutocompleteHandler.cpp
@@ -1,5 +1,0 @@
-//
-// Created by rosie on 21/10/2024.
-//
-
-#include "AutocompleteHandler.h"
